Refactored for adding the MOVE controller
